+++ package.json
@@ -14,6 +14,7 @@
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
     "react-konva": "^16.13.0-3",
+    "react-router-dom": "^5.2.0",
     "react-scripts": "3.4.1",
     "use-image": "^1.0.5"
   },
